+++ iAi_Project/iAi_Project/source/BulletEnemyRangeBeam.cpp
@@ -1,31 +1,31 @@
 /* 2025.01.29 石川智也 ファイル作成 */
 
-#include "BulletEnemyRangeNormal.h"
+#include "BulletEnemyRangeBeam.h"
 
-/* ノーマル弾クラスの定義 */
+/* ビームクラスの定義 */
 
 // コンストラクタ
-BulletEnemyRangeNormal::BulletEnemyRangeNormal() : BulletBase()
+BulletEnemyRangeBeam::BulletEnemyRangeBeam() : BulletBase()
 {
 	/* 初期化 */
 	this->iObjectType = OBJECT_TYPE_BULLET_ENEMY;	// オブジェクトの種類を"弾(エネミー)"に設定
 	this->pEffect = nullptr;
 
-	this->iDurationCount = ENEMY_NORMAL_DURATION_COUNT;		// 弾の持続カウント
-	this->iBulletCount = ENEMY_NORMAL_BULLET_COUNT;			// 弾発射カウント
+	this->iDurationCount = ENEMY_NORMAL_DURATION_COUNT;		// ビームの持続カウント
+	this->iBulletCount = ENEMY_NORMAL_BULLET_COUNT;			// ビーム発射カウント
 
-	this->iEnemyNormalDurationCount = ENEMY_NORMAL_DURATION_COUNT;	//ノーマル弾の持続カウント
+	this->iEnemyBeamDurationCount = ENEMY_NORMAL_DURATION_COUNT;	//ビームの持続カウント
 }
 
 // デストラクタ
-BulletEnemyRangeNormal::~BulletEnemyRangeNormal()
+BulletEnemyRangeBeam::~BulletEnemyRangeBeam()
 {
 	/* 紐づいているエフェクトの削除フラグを有効化 */
 	this->pEffect->SetDeleteFlg(true);
 }
 
 // 初期化
-void BulletEnemyRangeNormal::Initialization()
+void BulletEnemyRangeBeam::Initialization()
 {
 	/* 当たり判定設定 */
 	{
@@ -35,7 +35,7 @@ void BulletEnemyRangeNormal::Initialization()
 
 	/* エフェクト追加 */
 	{
-		/* ノーマル弾エフェクトを生成 */
+		/* ビームエフェクトを生成 */
 		this->pEffect = new EffectManualDelete();
 
 		/* エフェクトの読み込み */
@@ -60,15 +60,15 @@ void BulletEnemyRangeNormal::Initialization()
 	}
 }
 
-// ノーマル弾の移動処理
-void BulletEnemyRangeNormal::BulletEnemyRangeNormalMove()
+// ビームの移動処理
+void BulletEnemyRangeBeam::BulletEnemyRangeBeamMove()
 {
 	/* プレイヤーの座標を取得 */
 	CharacterBase* player = dynamic_cast<DataList_Object*>(gpDataListServer->GetDataList("DataList_Object"))->GetCharacterPlayer();
 	VECTOR playerPos = player->vecGetPosition();
 
 	// 持続カウントが発射カウントを超えているか確認
-	if (iEnemyNormalDurationCount >= ENEMY_NORMAL_BULLET_COUNT)
+	if (iEnemyBeamDurationCount >= ENEMY_NORMAL_BULLET_COUNT)
 	{
 		//持続カウントが発射カウントを超えている場合
 		// プレイヤーのy座標を取得
@@ -78,26 +78,27 @@ void BulletEnemyRangeNormal::BulletEnemyRangeNormalMove()
 		this->vecDirection = VNorm(VSub(playerPos, this->vecPosition));
 	}
 
-	// ノーマル弾の移動座標と向きと速度を更新
+	// ビームの移動座標と向きと速度を更新
 	this->vecPosition = VAdd(this->vecPosition, VScale(this->vecDirection, this->fMoveSpeed = 18));
 
-	// ノーマル弾のコリジョン座標を更新
+	// ビームのコリジョン座標を更新
 	this->stCollisionSqhere.vecSqhere = this->vecPosition;
 
-	// ノーマル弾のエフェクト座標を更新
+	// ビームのエフェクト座標を更新
 	this->pEffect->SetPosition(this->vecPosition);
+
 }
 
 // 更新
-void BulletEnemyRangeNormal::Update()
+void BulletEnemyRangeBeam::Update()
 {
 
 	// 持続カウントが0より大きいか確認
-	if (iEnemyNormalDurationCount > 0)
+	if (iEnemyBeamDurationCount > 0)
 	{
 		// 持続カウントが0より大きい場合
 		// 持続カウントを減算
-		iEnemyNormalDurationCount--;
+		iEnemyBeamDurationCount--;
 	}
 	else
 	{
@@ -106,6 +107,6 @@ void BulletEnemyRangeNormal::Update()
 		this->bDeleteFlg = true;
 	}
 
-	// ノーマル弾の移動処理
-	BulletEnemyRangeNormalMove();
+	// ビームの移動処理
+	BulletEnemyRangeBeamMove();
 }
